Enabled import for teachers. This is a naive import that ignoes some important features.
The teacher model does not yes support status, i.e. has no concept of what to do when a teacher leaves.
They cannot simply be deleted because this would invalidate any recordsthat they were linked with.
Instead they should have a status of archived or left or retired, somehting that enables them still to be on the database but shows them as not current.