A bit more about views. Specifically problems when calling an update view via a form button where no data appears in the form. Whereas when I call the form using the same URL from a A link the form appears with all the right data.
I suspect this has to do with GET and POST I need to look into it more.

This is not a problem iwth URLS as they appear to work fine it seemsto be a problem with GET sending all the data across and POST not doing so.

Conversely You can't call a generic deleteview from an A link, you get Django errors but you can call it from a POST without errors.